removed quota schema and relevant apis
