+++ fabric_cf/actor/core/apis/abc_database.py
@@ -478,43 +478,3 @@ class ABCDatabase(ABC):
         @param states states
         @param last_update_time last update time
         """
-
-    def create_quota(self, project_id: str, resource_type: str, resource_unit: str, quota_limit: int):
-        """
-        Create a new quota record in the database.
-
-        @param project_id: UUID of the project the quota is associated with.
-        @param resource_type: Type of resource (e.g., SLICE, COMPONENT).
-        @param resource_unit: Unit of the resource (e.g., HOURS, COUNT, GB).
-        @param quota_limit: Maximum allowed usage for this resource.
-        @return: The created `Quotas` object.
-        @throws: Exception if there is an error during the creation.
-        """
-
-    def get_quota_lookup(self, project_id: str):
-        """
-        Fetches all quotas for a given project and creates a lookup dictionary.
-
-        @param project_id: UUID of the project whose quotas are to be fetched.
-        @return: Dictionary with keys as (resource_type, resource_unit) and values as quota details.
-        @throws: Exception if there is an error during the database interaction.
-        """
-
-    def update_quota(self, reservation: ABCReservationMixin):
-        """
-        Update an existing quota record.
-
-        @param reservation: reservation.
-        @throws: Exception if there is an error during the update.
-        """
-
-    def delete_quota(self, project_id: str, resource_type: str, resource_unit: str):
-        """
-        Delete a specific quota record.
-
-        @param project_id: UUID of the project the quota is associated with.
-        @param resource_type: Type of resource (e.g., SLICE, COMPONENT).
-        @param resource_unit: Unit of the resource (e.g., HOURS, COUNT, GB).
-        @return: True if the quota was successfully deleted, False if not found.
-        @throws: Exception if there is an error during the deletion.
-        """
